Add default dir as git safe directory
Git introduced the usage of `safe.directory` in response to the security
vulnerability CVE-2022-24765. When running `ansible-lint`, the directory
mounted to `/data` is often not root-owned, but `/data is, so we get
errors:

Failed to guess project directory using git: fatal: detected dubious
ownership in repository at '/data'
To add an exception for this directory, call:
	git config --global --add safe.directory /data
WARNING  Failed to discover lintable files using git: fatal: detected
dubious ownership in repository at '/data'

This commit creates a `/root/.gitconfig` including `/data` as safe
directory.